+++ CMakeLists.txt
@@ -1,31 +1,27 @@
-cmake_minimum_required(VERSION 3.5)
+cmake_minimum_required(VERSION 3.13)
 
-project(Livewire VERSION 0.1 LANGUAGES CXX)
+project(Livewire
+    VERSION 0.1
+    LANGUAGES CXX)
 
-set(CMAKE_INCLUDE_CURRENT_DIR ON)
-
-set(CMAKE_AUTOUIC ON)
-set(CMAKE_AUTOMOC ON)
-set(CMAKE_AUTORCC ON)
-
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-enable_testing()
-
-find_package(QT NAMES Qt6 Qt5 REQUIRED COMPONENTS Widgets)
-find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Widgets)
-find_package(OpenCV REQUIRED)
+set(CMAKE_CXX_EXTENSIONS OFF)
 
 Include(FetchContent)
 
-FetchContent_Declare(
-    Catch2
-    GIT_REPOSITORY https://github.com/catchorg/Catch2.git
-    GIT_TAG        v3.0.1
-)
+add_subdirectory(src/livewire)
+add_subdirectory(src/app)
 
-FetchContent_MakeAvailable(Catch2)
+string(
+        COMPARE EQUAL
+        "${CMAKE_SOURCE_DIR}"
+        "${PROJECT_SOURCE_DIR}"
+        PROJECT_IS_TOP_LEVEL)
 
-add_subdirectory(src)
-add_subdirectory(test)
+option(ENABLE_TESTING "Enable the tests" ${PROJECT_IS_TOP_LEVEL})
+if(ENABLE_TESTING)
+    include(CTest)
+    enable_testing()
+    add_subdirectory(test)
+endif()
