+++ src/app/paint_area.cpp
