+++ src/app/paint_area.hpp
