+++ test/CMakeLists.txt
@@ -1,3 +1,11 @@
+FetchContent_Declare(
+    Catch2
+    GIT_REPOSITORY https://github.com/catchorg/Catch2.git
+    GIT_TAG        v3.0.1
+)
+
+FetchContent_MakeAvailable(Catch2)
+
 add_executable(tests test_main.cpp)
 target_link_libraries(tests PRIVATE Catch2::Catch2WithMain)
 add_test(my_tests tests)
